# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: LogPulse API
  version: 1.1.0
  description: |
    REST API for LogPulse — a centralized log monitoring platform.

    Three authentication realms exist:

    | Realm | Mechanism | Used by |
    |-------|-----------|---------|
    | Dashboard JWT | `Authorization: Bearer <jwt>` (24h expiry, HS256) | everything under `/api/**` except ingest |
    | Ingest API key | `Authorization: Bearer <application API key>` | `POST /api/logs` only |
    | OAuth redirect | browser 302 flow | `/api/auth/{google,github}/*` |

    There is no public signup: an admin invites users, and the invited person
    activates the account via `POST /api/auth/accept-invite` or by signing in
    with Google/GitHub.

    Errors are always `{"error": "<human readable message>"}`.
servers:
  - url: http://localhost:8080
    description: Local development
  # INFERRED: production URL depends on your deployment; set your own.
  - url: https://your-logpulse-host
    description: Production (replace with your host)
...
paths:
...
  /api/logs:
    post:
      tags: [Logs]
      summary: Ingest a log entry
      description: >
        Authenticated by the **application API key** (not a user JWT). The
        message is fingerprinted at ingest (variables normalized) so the
        clustering endpoints can group near-duplicates. Levels are
        normalized: `warn`→`warning`, `err`→`error`, `info`→`info`,
        `crit`→`critical`; any other value is stored as-is.
      operationId: ingestLog
      security: [{ ingestApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_name, level, message]
              properties:
                app_name: { type: string, example: payments-api }
                level:
                  type: string
                  description: 'info | warning | error | critical | debug (normalized)'
                  example: error
                message: { type: string }
                meta:
                  type: object
                  additionalProperties: true
                  description: Free-form structured fields; become queryable evidence.
      responses:
        '201':
          description: Stored (and broadcast over /ws/logs)
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, example: stored }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Missing or invalid API key
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '500':
          description: MongoDB write failed
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      tags: [Logs]
      summary: Search & filter logs
      description: >
        Returns the newest 200 matching entries (timestamp descending). All
        filters are optional and combined with AND.
      operationId: listLogs
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: app, in: query, schema: { type: string }, description: Exact app_name match }
        - { name: level, in: query, schema: { type: string }, description: Exact level match }
        - { name: q, in: query, schema: { type: string }, description: Case-insensitive substring match on message }
      responses:
        '200':
          description: Matching log entries
          content:
            application/json:
              schema:
                type: array
                maxItems: 200
                items: { $ref: '#/components/schemas/LogEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # -------------------------------------------------- Stats & Monitoring ----
...
  /api/incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      description: Newest first, max 200. Without a status filter, all statuses are returned.
      operationId: listIncidents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, open, acknowledged, resolved] }
          description: '`active` = open + acknowledged.'
      responses:
        '200':
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Incident' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Incidents]
      summary: Declare an incident manually
      operationId: createIncident
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                severity: { type: string, enum: [critical, error, warning, info], default: error }
                affected_apps:
                  type: string
                  description: JSON-encoded array of app names, e.g. '["api","worker"]'
                summary: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/applications:
    get:
      tags: [Applications]
      summary: List registered applications (with API keys)
      operationId: listApplications
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Applications, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Application' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Applications]
      summary: Register an application
      description: Generates a fresh API key (crypto/rand, 64 hex chars).
      operationId: createApplication
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Created — store the api_key now; it is shown in the list later too
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------- Saved Searches ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard JWT from /api/auth/login — 24h expiry (HS256).
    ingestApiKey:
      type: http
      scheme: bearer
      description: Per-application API key from POST /api/applications.
...
  responses:
    ValidationError:
      description: Request body failed validation (message is the validator error)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing/expired JWT ("missing token" / "invalid or expired token")
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    LogEntry:
      type: object
      properties:
        id: { type: string, description: Mongo ObjectId }
        app_name: { type: string }
        level:
          type: string
          description: >
            Normalized at ingest (warn→warning, err→error, crit→critical);
            arbitrary values pass through.
        message: { type: string }
        meta: { type: object, additionalProperties: true, nullable: true }
        timestamp: { type: string, format: date-time }
        pattern_hash:
          type: string
          nullable: true
          description: First 16 hex chars of the SHA-256 of the normalized pattern (AI-1)
        pattern:
          type: string
          nullable: true
          description: Normalized message template, e.g. "failed to charge card <id> <id>"
...
    Incident:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        severity: { type: string, enum: [critical, error, warning, info] }
        status: { type: string, enum: [open, acknowledged, resolved] }
        source: { type: string, enum: [anomaly, correlation, alert, manual] }
        summary: { type: string }
        evidence: { type: string, description: JSON-encoded array (reserved for stored evidence) }
        affected_apps: { type: string, description: JSON-encoded array of app names }
        alert_event_ids: { type: string, description: JSON-encoded array of linked AlertEvent ids (AI-6) }
        started_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        created_by_user_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Application:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        api_key: { type: string, description: Bearer key for POST /api/logs }
        owner_id: { type: integer }
        created_at: { type: string, format: date-time }
...